[JIT] Specialize autograd zero: fix the guarding condition.
We are looking for tensors that are expected to be undefined (according
to the profile info) and should be checking for them to satisfy the
following condition: "not(have any non-zero)", which is equivalent to
"tensor is all zeros". The issue was that we've been checking tensors
that were expected *not* to be undefined.

ghstack-source-id: 4946fdbddaf899d86278af68e2e84e1bda06bae4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43846