Switch all Sequences in tools.codegen.model to Tuple
I thought I was being clever by using Sequence, which doesn't commit to
List or Tuple, but forces read-onlyness in the type system.  However,
there is runtime implication to using List or Tuple: Lists can't be
hashed, but Tuples can be!  This is important because I shortly want
to group by FunctionSchema, and to do this I need FunctionSchema to
be hashable.  Switch everything to Tuple for true immutability.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 6dc30936fa42f5e6ef3b45db9f08c45956bcef5b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45127