[RFC] Switch over gen.py to using the selective build abstraction instead of directly querying op_registration_whitelist
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45211

Currently, the selective build machinery queries `op_registration_whitelist` directly at various places in the code. `op_registration_whitelist` is a list of allowed operator names (without overload name). We want to move to a world where the overload names are also included so that we can be more selective about which operators we include. To that effect, it makes sense to hide the checking logic in a separate abstraction and have the build use that abstraction instead of putting all this selective build specific logic in the code-generator itself. This change is attempting to do just that.

I have explicitly not touched the part about per-op-registration since I have another change set out to remove it entirely since it's not used, so I didn't think it would be worth the time spent making the change and then resolving the merge conflict.
ghstack-source-id: 112736377

Differential Revision: [D23872538](https://our.internmc.facebook.com/intern/diff/D23872538/)