[JIT] Enable @unused syntax for ignoring properties
**Summary**
This commit enables `@unused` syntax for ignoring
properties. Inoring properties is more intuitive with this feature enabled.
`@ignore` is not supported because class type properties cannot be
executed in Python (because they exist only as TorchScript types) like
an `@ignored` function and module properties that cannot be scripted
are not added to the `ScriptModule` wrapper so that they
may execute in Python.

**Test Plan**
This commit updates the existing unit tests for class type and module
properties to test properties ignored using `@unused`.

ghstack-source-id: 5397644794fe1c4c7577090e6bedfe36964c6e70
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45261